--- lib/presentation/bloc/diary_event.dart
@@ -1,13 +1,0 @@
-// part of 'diary_bloc.dart';
-
-// @immutable
-// abstract class DiaryEvent {}
-
-// class InitialEnvent extends DiaryEvent {}
-
-// class AddDiaryCard extends DiaryEvent {
-//   final String title;
-//   final String desc;
-
-//   AddDiaryCard(this.title, this.desc);
-// }
