--- lib/presentation/bloc/diary_state.dart
@@ -1,26 +1,0 @@
-// part of 'diary_bloc.dart';
-
-// @immutable
-// abstract class DiaryState {}
-
-// class DiaryInitial extends DiaryState {}
-
-// class DiaryInitialLoaded extends DiaryState {
-//   final Stream<List<DiaryEntry>> diaryEntryStream;
-
-//   DiaryInitialLoaded(this.diaryEntryStream);
-// }
-
-// class DiaryLoading extends DiaryState {}
-
-// class DiaryCardAdded extends DiaryState {
-//   final List<DiaryEntry> diaryCardDetailsList;
-
-//   DiaryCardAdded(this.diaryCardDetailsList);
-// }
-
-// class DiaryError extends DiaryState {
-//   final String errorMessage;
-
-//   DiaryError(this.errorMessage);
-// }
